fix lint issues
```release-note
NONE
```
