+++ app/plugin_api_tests/test_db_driver/main.go
@@ -41,12 +41,13 @@ func (p *MyPlugin) MessageWillBePosted(_ *plugin.Context, _ *model.Post) (*model
 		store.GetMaster().Db = handle
 		store.SetMasterX(handle)
 
+		wrapper := sqlstore.NewStoreTestWrapper(store)
 		// Testing with a handful of stores
-		storetest.TestPostStore(p.t, store, store)
-		storetest.TestUserStore(p.t, store, store)
+		storetest.TestPostStore(p.t, store, wrapper)
+		storetest.TestUserStore(p.t, store, wrapper)
 		storetest.TestTeamStore(p.t, store)
-		storetest.TestChannelStore(p.t, store, store)
-		storetest.TestBotStore(p.t, store, store)
+		storetest.TestChannelStore(p.t, store, wrapper)
+		storetest.TestBotStore(p.t, store, wrapper)
 
 		store.GetMaster().Db.Close()
 	}
